feat/app_main_script_refactor (#9)
* feat: sorting function calls according to main script sequence for readability

* feat: updating and cleaning according to last refactors

* fix: cleaning and re-structuring according with refactors

* fix: correcting folder naming to standardize

* feat: refactoring helm manager v1 modules

* feat: refactoring create_vars_file to modular structure

* feat: refactoring create_environment_values_file to module structure

* feat: refactoring to_dict to modular

* feat: refactoring read_env to modular kubernetes services

* feat: refactoring read_env_from to module kubernetes_services

* feat: refactoring read_volumes to module kubernetes_services

* feat: refactoring read_volumes_mounts to module kubernetes_services

* feat: refactoring read_annotations to module kubernetes_services

* feat: refactoring read_ingress_rules to module kubernetes_services

* feat: refactoring read_ingress_tls to kubernetes_services module

* feat: refactoring read_image_pull_secrets to kubernetes_service module

* feat: refactoring read_host_aliases to kubernetes_services module

* feat: refactoring create_services to kubernetes_services module

* feat: refactoring create_workload_template to kubernetes_manager module

* fix: correcting directory name for better definition

* feat: refactoring create_configmap_or_secret to kubernetes_manager module

* feat: refactoring create_ingress to kubernetes_manager module

* feat: refactoring load_kubernetes_config to kubernetes_manager module

* feat: refactoring load_kubernetes_data to kubernetes_manager module
