+++ app/v1/modules/kubernetes_manager/script_create_vars_file.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+
+
+# Application
+from app.v1.modules.file_manager import ScriptFileWriterManager
+from app.v1.modules.utility_services import ScriptUtilDictCleaner
+
+
+class ScriptVarsFileCreator:
+    """
+    ScriptVarsFileCreator
+    """
+
+    @staticmethod
+    def create_vars_file(conf):
+        field = dict(
+            ConfigMap='data',
+            Secret='stringData'
+        )
+        for kind in conf['kubernetes']['values']:
+            if kind in ['ConfigMap', 'Secret']:
+                for item in conf['kubernetes']['values'][kind]:
+                    ScriptFileWriterManager.write_file(
+                        f"config_files/output/vars/{conf['chart']['name']}/{item}.json",
+                        ScriptUtilDictCleaner.remove_empty_from_dict(conf['kubernetes']['values'][kind][item][field[kind]]),
+                        mode='json'
+                    )
+        pass
