+++ app/v1/modules/utility_services/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .script_remove_empty_from_dict import ScriptUtilDictCleaner
+
+
+__all__ = [
+    "ScriptUtilDictCleaner",
+]
